Extract bandwidth and encoding related functions
